tests/wpt/tests/presentation-api/controlling-ua/PresentationRequest_success.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Constructing a PresentationRequest</title>
<link rel="author" title="Franck William Taffo" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs/">
<link rel="help" href="http://w3c.github.io/presentation-api/#constructing-a-presentationrequest">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  test(() => {
    let request = new PresentationRequest('presentation.html');
    assert_true(request instanceof PresentationRequest, 'An instance of PresentationRequest with a relative presentation URL is constructed successfully.');

    request = new PresentationRequest('https://example.org/');
    assert_true(request instanceof PresentationRequest, 'An instance of PresentationRequest with an absolute presentation URL is constructed successfully.');

    request = new PresentationRequest([
      'presentation.html',
      'https://example.org/presentation/'
    ]);
    assert_true(request instanceof PresentationRequest, 'An instance of PresentationRequest with an array of presentation URLs is constructed successfully.');

    request = new PresentationRequest([
      'unsupported://example.com',
      'presentation.html',
      'https://example.org/presentation/'
    ]);
    assert_true(request instanceof PresentationRequest, 'An unsupported URL in an array of presentation URLs is ignored successfully.');
  });
</script>
